Bring graph compression down to Cython
Currently this yields an approximate 2.7x increase in performance (10s down to 3.7 on a 5600x) when using the Long An
model. More testing and proper benchmarking is to come.

I've looked into parallelising this code but if done naively a thread local `slink` variable, acting as a counter,
would provide random access to the `compression_.*` arrays. I'm not exactly sure how to move forward with parallelising
it without some more invasive changes.

Some clean up commits are to come regarding the setting of members which is currently done at the end of the cythonised
funciton.